feat: 통계 조회 시 year 파라미터가 없다면 전체 기간을 조회하도록 기능 추가

+++ backend/app/src/main/java/com/yagubogu/stat/dto/StadiumStatsParam.java
@@ -2,7 +2,7 @@ package com.yagubogu.stat.dto;
 
 public record StadiumStatsParam(
         String stadiumName,
-        long winCounts,
-        long totalCountsWithoutDraw
+        int winCounts,
+        int totalCountsWithoutDraw
 ) {
 }
